(feat) 修改更新地址和删除地址的bug

+++ services/users/internal/logic/getaddresslogic.go
@@ -30,7 +30,7 @@ func NewGetAddressLogic(ctx context.Context, svcCtx *svc.ServiceContext) *GetAdd
 func (l *GetAddressLogic) GetAddress(in *users.GetAddressRequest) (*users.GetAddressResponse, error) {
 	// todo: add your logic here and delete this line
 
-	address, err := l.svcCtx.AddressModel.GetUserAddressbyIdAndUserId(l.ctx, in.AddressId, in.UserId)
+	address, err := l.svcCtx.AddressModel.GetUserAddressbyIdAndUserId(l.ctx, in.AddressId, int32(in.UserId))
 	if err != nil {
 		if errors.Is(err, sql.ErrNoRows) {
 			l.Logger.Infow("get address not found", logx.Field("user_id", in.UserId), logx.Field("address_id", in.AddressId))
